requirements.txt
django~=4.2
djangorestframework~=3.14.0
psycopg2-binary
python-decouple
django-filter
dj_database_url
pillow
drf-yasg
djangorestframework-simplejwt
celery~=5.2.7
django-celery-beat
django-celery-results
redis

dj-config-url~=0.1.1
decouple~=0.0.7